services/course_management/requirement.txt
anyio==3.6.2
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
deprecation==2.1.0
ecdsa==0.18.0
fastapi==0.95.1
h11==0.14.0
idna==3.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
python-dotenv==1.0.0
python-jose==3.3.0
python-keycloak==2.16.2
requests==2.30.0
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tomli==2.0.1
typing-extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
